requirements.txt
Flask==3.0.3
torch==2.5.1
torchvision==0.20.1
Pillow==10.2.0
transformers==4.46.2
werkzeug==3.1.1